commit create

+++ diplomatiki/src/main/java/diplomatiki/Repositories/ProfessorRepository.java
@@ -6,6 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProfessorRepository extends JpaRepository<Professor, Long> {
+public interface ProfessorRepository extends JpaRepository<Professor, Integer> {
+
+    // Για login: βρίσκει καθηγητή από User
     Optional<Professor> findByUser(User user);
 }
